+++ src/StreetNaming.Web/Migrations/StreetNamingEntitiesModelSnapshot.cs
@@ -17,7 +17,7 @@ namespace StreetNaming.Web.Migrations
 
             modelBuilder.Entity("StreetNaming.Domain.Models.Applicant", b =>
                 {
-                    b.Property<long>("ApplicantId")
+                    b.Property<int>("ApplicantId")
                         .ValueGeneratedOnAdd();
 
                     b.Property<string>("County");
@@ -75,13 +75,13 @@ namespace StreetNaming.Web.Migrations
 
             modelBuilder.Entity("StreetNaming.Domain.Models.Attachment", b =>
                 {
-                    b.Property<long>("AttachmentId")
+                    b.Property<int>("AttachmentId")
                         .ValueGeneratedOnAdd();
 
                     b.Property<byte[]>("Bytes")
                         .IsRequired();
 
-                    b.Property<long>("CaseId");
+                    b.Property<int>("CaseId");
 
                     b.Property<string>("ContentType")
                         .IsRequired()
@@ -104,13 +104,13 @@ namespace StreetNaming.Web.Migrations
 
             modelBuilder.Entity("StreetNaming.Domain.Models.Case", b =>
                 {
-                    b.Property<long>("CaseId")
+                    b.Property<int>("CaseId")
                         .ValueGeneratedOnAdd();
 
                     b.Property<string>("AdditionalInformation")
                         .HasAnnotation("MaxLength", 2000);
 
-                    b.Property<long>("ApplicantId");
+                    b.Property<int>("ApplicantId");
 
                     b.Property<int>("CaseStatus");
 
@@ -125,7 +125,7 @@ namespace StreetNaming.Web.Migrations
 
                     b.Property<DateTime?>("EffectiveDate");
 
-                    b.Property<long?>("ExistingPropertyUrn");
+                    b.Property<int?>("ExistingPropertyUrn");
 
                     b.Property<bool>("IsRegisteredOwner");
 
@@ -161,12 +161,12 @@ namespace StreetNaming.Web.Migrations
 
             modelBuilder.Entity("StreetNaming.Domain.Models.Transaction", b =>
                 {
-                    b.Property<long>("TransactionId")
+                    b.Property<int>("TransactionId")
                         .ValueGeneratedOnAdd();
 
                     b.Property<decimal>("Amount");
 
-                    b.Property<long>("CaseId");
+                    b.Property<int>("CaseId");
 
                     b.Property<DateTime>("CreatedDate")
                         .ValueGeneratedOnAdd()
